clean inspections in functions created utils package to separate functions created and fixed reads.py
